Show media buttons in overlay
